Renamed files, classes and folders. Added NuGet packages to run tests in VS2019
