+++ Source/XtensionSuite/XtensionSuite.csproj
@@ -78,9 +78,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="BDDXtensions\BoolBDDXtensions.cs" />
-    <Compile Include="BDDXtensions\NumericBDDXtensions.cs" />
-    <Compile Include="BDDXtensions\StringBDDXtensions.cs" />
+    <Compile Include="AssertXtensions\BoolAssertXtensions.cs" />
+    <Compile Include="AssertXtensions\NumericAssertXtensions.cs" />
+    <Compile Include="AssertXtensions\StringAssertXtensions.cs" />
     <Compile Include="BoolXtensions\BoolXtensions.cs" />
     <Compile Include="DateTimeXtensions\DateTimeXtensions_BasedOnAge.cs" />
     <Compile Include="DateTimeXtensions\DateTimeXtensions_BasedOnDays.cs" />
@@ -132,12 +132,6 @@
       <Install>true</Install>
     </BootstrapperPackage>
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="DictionaryXtensions" />
-    <Folder Include="ListXtensions" />
-    <Folder Include="GenericXtensions" />
-    <Folder Include="TimeSpanXtensions" />
-    <Folder Include="ObjectXtensions" />
-  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.Targets" />
 </Project>
