+++ Source/XtensionSuiteTests/BoolXtensionsTests.cs
@@ -6,8 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using NUnit.Framework;
-using Nallenthal.Common.XtensionSuite.BDDExtensions.BoolBDDXtensions;
-using Nallenthal.Common.XtensionSuite.BDDExtensions.NumericBDDXtensions;
+using Nallenthal.Common.XtensionSuite.AssertExtensions;
 using Nallenthal.Common.XtensionSuite.BoolXtensions;
 
 namespace Nallenthal.Common.XtensionSuiteTests
