+++ Source/XtensionSuiteTests/packages.config
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="NUnit" version="3.11.0" targetFramework="net40" />
-  <package id="NUnit3TestAdapter" version="3.13.0" targetFramework="net40" />
+  <package id="NUnit" version="3.13.3" targetFramework="net40" />
+  <package id="NUnit3TestAdapter" version="4.3.1" targetFramework="net40" />
 </packages>
